+++ etc/dbus-serialbattery/bms/jkbms_ble.py
@@ -1,4 +1,10 @@
 # -*- coding: utf-8 -*-
+
+# Notes
+# Added by https://github.com/baranator
+# https://github.com/Louisvdw/dbus-serialbattery/pull/372
+# Updated by https://github.com/mr-manuel
+
 from battery import Battery, Cell
 from typing import Callable
 from utils import logger, AUTO_RESET_SOC