+++ etc/dbus-serialbattery/bms/mnb_test_max17853.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 import spidev
 import time
 import math